Add SocialIcons component
